refactor: flatten symbology folder structure
Move grammar compiler files and components to the top-level symbology
directory, eliminating the now-unnecessary grammar/ and vector_layer/
subdirectories.

- grammar/grammarToOLLayer.ts → grammarToOLLayer.ts
- grammar/grammarToOLStyle.ts → grammarToOLStyle.ts
- grammar/components/{MappingRow,ScaleEditor,NumericInput}.tsx → components/
- vector_layer/types/Grammar.tsx → Grammar.tsx
- vector_layer/types/Heatmap.tsx → Heatmap.tsx
- Delete VectorRendering.tsx (inline dispatch into symbologyDialog)
- Delete grammar/types.ts (re-exports only; importers now use @jupytergis/schema)
- Remove accidentally committed project1.qgz
- Fix InterpolationType: define inline in classificationModes.ts instead
  of importing from the deleted SingleBandPseudoColor
- Fix stale jest.mock paths in test files
- Add type="button" to all <button> elements in symbology dialog components
  to prevent accidental form submission in JupyterLab's Dialog (which wraps
  its body in a <form>)
