+++ packages/base/src/features/layers/symbology/components/NumericInput.tsx
