--- packages/base/src/features/layers/symbology/grammar/types.ts
@@ -1,28 +1,0 @@
-// Re-exported from @jupytergis/schema — the canonical location for these types.
-export type {
-  RGBA,
-  ClassificationMode,
-  ICompareOp,
-  IPredicate,
-  OutputType,
-  RGBAChannel,
-  UInt8Channel,
-  UNormChannel,
-  PosFloatChannel,
-  StyleChannel,
-  IColorRampScale,
-  ICategoricalScale,
-  IScalarScale,
-  IConstantRGBAScale,
-  IConstantNumScale,
-  IIdentityScale,
-  IExpressionScale,
-  IScale,
-  IMapping,
-  IEncodingRule,
-  ITransform,
-  IKDETransform,
-  IClusterTransform,
-  IGrammarLayer,
-  IGrammarSymbologyState,
-} from '@jupytergis/schema';
