+++ packages/base/src/features/layers/symbology/symbologyDialog.tsx
@@ -6,8 +6,8 @@ import { Signal } from '@lumino/signaling';
 import React, { useEffect, useState } from 'react';
 
 import { SymbologyTab, SymbologyValue } from '@/src/types';
-import VectorRendering from './vector_layer/VectorRendering';
-import Grammar from './vector_layer/types/Grammar';
+import Grammar from './Grammar';
+import Heatmap from './Heatmap';
 
 export interface ISymbologyDialogProps {
   model: IJupyterGISModel;
@@ -87,11 +87,9 @@ const SymbologyDialog: React.FC<ISymbologyDialogProps> = ({
 
     // TODO GeoTiffLayers can also be used for other layers, need a better way to determine source + layer combo
     switch (layer.type) {
-      case 'VectorLayer':
-      case 'VectorTileLayer':
       case 'HeatmapLayer':
         LayerSymbology = (
-          <VectorRendering
+          <Heatmap
             model={model}
             okSignalPromise={okSignalPromise}
             layerId={selectedLayer}
@@ -100,6 +98,8 @@ const SymbologyDialog: React.FC<ISymbologyDialogProps> = ({
           />
         );
         break;
+      case 'VectorLayer':
+      case 'VectorTileLayer':
       case 'GeoTiffLayer':
         LayerSymbology = (
           <Grammar
